external/requirements.txt
lxml
requests
progressbar2